api done

+++ server/controllers/task.controller.js
@@ -1,11 +1,12 @@
 import asyncHandler from 'express-async-handler';
 
 import Task from '../models/task.model.js';
+import User from '../models/user.model.js';
 
 // @des		GET all tasks
 // @route   GET /api/v1/tasks
 export const getAllTasks = asyncHandler(async (req, res) => {
-   const tasks = await Task.find();
+   const tasks = await Task.find({ user: req.user.id });
    res.status(200).json({ success: true, data: tasks });
 });
 
@@ -23,11 +24,14 @@ export const getTask = asyncHandler(async (req, res) => {
 // @des	   CREATE a new task
 // @route   POST /api/v1/tasks
 export const createTask = asyncHandler(async (req, res) => {
-   if (!req.body.taskName) {
+   if (!req.body.name) {
       res.status(400);
       throw new Error('Please enter a new task name');
    }
-   const newTask = await Task.create({ taskName: req.body.taskName });
+   const newTask = await Task.create({
+      user: req.user.id,
+      name: req.body.name,
+   });
    res.status(201).json({
       success: true,
       data: newTask,
@@ -42,6 +46,18 @@ export const updateTask = asyncHandler(async (req, res) => {
       res.status(404);
       throw new Error(`Task with id ${req.params.id} not found`);
    }
+
+   const user = await User.findById(req.user.id);
+   if (!user) {
+      res.status(401);
+      throw new Error('User not found');
+   }
+
+   if (String(task.user) !== user.id) {
+      res.status(401);
+      throw new Error('User not authorized');
+   }
+
    const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
    });
@@ -56,6 +72,18 @@ export const deleteTask = asyncHandler(async (req, res) => {
       res.status(404);
       throw new Error(`Task with id ${req.params.id} not found`);
    }
+
+   const user = await User.findById(req.user.id);
+   if (!user) {
+      res.status(401);
+      throw new Error('User not found');
+   }
+
+   if (String(task.user) !== user.id) {
+      res.status(401);
+      throw new Error('User not authorized');
+   }
+
    const deletedTask = await Task.findByIdAndDelete(req.params.id);
    res.status(200).json({ success: true, data: deletedTask });
 });
